# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Warehouse Management API
  version: 1.0.0
  description: |
    Complete REST API contract for Warehouse Management backend Phases 1–7.
    All dates are ISO-8601; date-only report endDate values include that entire
    UTC calendar day. Private endpoints require a Bearer access token.
servers:
  - url: /api/v1
    description: Version 1 API
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users
      description: ADMIN only. Supports search/status and pagination.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - { name: search, in: query, schema: { type: string }, description: Search names/email }
        - { name: status, in: query, schema: { type: string, enum: [active, inactive, all] } }
      responses:
        "200": { description: Paginated safe user records }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Users]
      summary: Create an internal user
      description: ADMIN only; public self-registration is not available.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [firstName, lastName, email, password, role]
              properties:
                firstName: { type: string }
                lastName: { type: string }
                email: { type: string, format: email }
                password: { type: string, format: password }
                role: { $ref: "#/components/schemas/Role" }
                assignedWarehouses: { type: array, items: { $ref: "#/components/schemas/Id" } }
      responses:
        "201": { description: Created safe user record }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /transfers:
    get:
      tags: [Transfers]
      summary: List transfers visible to the current user
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [PENDING, APPROVED, COMPLETED, REJECTED, CANCELLED] } }
        - { name: product, in: query, schema: { $ref: "#/components/schemas/Id" } }
        - { name: warehouse, in: query, schema: { $ref: "#/components/schemas/Id" } }
        - { name: sourceWarehouse, in: query, schema: { $ref: "#/components/schemas/Id" } }
        - { name: destinationWarehouse, in: query, schema: { $ref: "#/components/schemas/Id" } }
        - { name: initiatedBy, in: query, schema: { $ref: "#/components/schemas/Id" } }
        - { name: reference, in: query, schema: { type: string } }
        - { name: startDate, in: query, schema: { type: string, format: date-time } }
        - { name: endDate, in: query, schema: { type: string, format: date-time } }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - { name: sort, in: query, schema: { type: string } }
      responses:
        "200": { description: Paginated visible transfers }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Transfers]
      summary: Request stock transfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId, quantity, sourceWarehouseId, sourceLocationId, destinationWarehouseId, destinationLocationId]
              properties:
                productId: { $ref: "#/components/schemas/Id" }
                quantity: { type: number, exclusiveMinimum: 0 }
                sourceWarehouseId: { $ref: "#/components/schemas/Id" }
                sourceLocationId: { $ref: "#/components/schemas/Id" }
                destinationWarehouseId: { $ref: "#/components/schemas/Id" }
                destinationLocationId: { $ref: "#/components/schemas/Id" }
                reference: { type: string }
                notes: { type: string }
      responses:
        "201": { description: PENDING transfer; stock is unchanged }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Id:
      type: string
      pattern: "^[a-fA-F0-9]{24}$"
      example: 507f1f77bcf86cd799439011
    Role:
      type: string
      enum: [ADMIN, MANAGER, STAFF]
    Error:
      type: object
      required: [success, message, data]
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
        data: { nullable: true, type: object, example: null }
        details: { type: object, additionalProperties: true }
...
  parameters:
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
...
  responses:
    ValidationError:
      description: Invalid input or filter.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, invalid, expired, or inactive-user authentication.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated user lacks the required role or warehouse scope.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Duplicate record, insufficient stock, or invalid state transition.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...